feat: default output used by Ninja

+++ test/13_cpp_config/config.cc
@@ -18,7 +18,7 @@ int main(int argc, const char** argv) {
   cc_project_t p = cc.project.create("cpp_config", CCProjectTypeConsoleApplication, NULL);
   // Files are added relative to the main CConstruct config file.
   std::vector<const char*> files{"src/main.c", "src/function.c", "src/function.h"};
-  cc.project.addFiles(p, files.size(), files.data(), NULL);
+  cc.project.addFiles(p, (unsigned int)files.size(), files.data(), NULL);
 
   cc_default_generator("build");
